+++ resources/components/laraknife/create-panel.blade.php
@@ -1,4 +1,4 @@
-@props(['title', 'error' => ''])
+@props(['title'])
 <div id="main-content" class="container mt-5">
     <x-laraknife.main-header title="{{ $title }}" />
     <fieldset class="lkn-create-panel">
@@ -7,6 +7,6 @@
             <x-laraknife.btn-store width1="2" width2="4" />
             <x-laraknife.btn-cancel width1="2" width2="4" />
         </div>
-        <x-laraknife.form-error error="{{$error}}" />
+        <x-laraknife.form-error />
     </fieldset>
 </div>