+++ resources/components/laraknife/field-error.blade.php
@@ -1,0 +1,5 @@
+@props(['name' => 'text'])
+@if ($errors->has($name))
+<div class="lkn-field-error">{{ $errors->first($name) }}</div>
+@endif
+