+++ resources/views/sproperty/index.blade.php
@@ -7,11 +7,11 @@
         <x-laraknife.filter-panel legend="{{ $pagination->legendText() }}">
             <x-laraknife.combobox position="first" name="scope" label="Scope" :options="$options"
                 width2="4" class="lkn-autoupdate" />
-            <x-laraknife.text position="last" name="text" label="Text" value="{{ $fields['text'] }}"
+            <x-laraknife.text position="last" name="text" label="Text" value="{{ $context->valueOf('text') }}"
                 width2="4" />
         </x-laraknife.filter-panel>
         <x-laraknife.index-button-panel buttonType="new"/>
-        <x-laraknife.sortable-table-panel :fields="$fields" :pagination="$pagination">
+        <x-laraknife.sortable-table-panel :context="$context" :pagination="$pagination">
             <thead>
                 <tr>
                     <th></th>